Mise en place d'un fichier specifique pour les fonctions purement lies a l'application Qt
